Extract HIR const function validation
Move const fn body and expression validation out of the HIR facade into hir/const_functions.rs, then lower the compiler source monolith ratchets for the smaller hir.rs footprint.

Validation: cargo fmt --manifest-path stage1/crates/axiomc/Cargo.toml --check; python3 scripts/ci/report-compiler-source-monoliths.py --json --check-plan --check-ratchet; make stage1-compiler-source-monoliths-test; make stage1-hir-boundary; make stage1-compiler-source-monoliths; cargo test --manifest-path stage1/Cargo.toml -p axiomc --lib hir::boundary_tests -- --nocapture; cargo test --manifest-path stage1/Cargo.toml -p axiomc --lib; git diff --check; git diff --cached --check.

Refs #1254.
