+++ stage1/crates/axiomc/src/hir.rs
@@ -7,6 +7,7 @@ use std::collections::{HashMap, HashSet};
 mod async_runtime;
 mod capabilities;
 mod const_arrays;
+mod const_functions;
 mod control_flow;
 mod definitions;
 mod diagnostics;
@@ -38,6 +39,7 @@ use self::const_arrays::{
     declared_array_len, eval_const_int_expr, resolve_const_int_decls,
     validate_const_array_lengths_in_program,
 };
+use self::const_functions::validate_const_function_body;
 use self::definitions::{
     VariantInfo, collect_enum_definitions, collect_struct_definitions, collect_trait_definitions,
     collect_type_names, validate_recursive_type_cycles, validate_trait_bounds_in_program,
@@ -407,119 +409,6 @@ fn lower_static_decls(
         });
     }
     Ok(lowered)
-}
-
-fn validate_const_function_body(
-    function: &syntax::Function,
-    functions: &HashMap<String, FunctionSig>,
-) -> Result<(), Diagnostic> {
-    if !function.is_const {
-        return Ok(());
-    }
-    if function.is_async {
-        return Err(Diagnostic::new(
-            "type",
-            format!("const fn {:?} cannot be async", function.name),
-        )
-        .with_span(function.line, function.column));
-    }
-    if function.is_extern {
-        return Err(Diagnostic::new(
-            "type",
-            format!("const fn {:?} cannot be extern", function.name),
-        )
-        .with_span(function.line, function.column));
-    }
-    for stmt in &function.body {
-        validate_const_function_stmt(function, stmt, functions)?;
-    }
-    Ok(())
-}
-
-fn validate_const_function_stmt(
-    function: &syntax::Function,
-    stmt: &syntax::Stmt,
-    functions: &HashMap<String, FunctionSig>,
-) -> Result<(), Diagnostic> {
-    match stmt {
-        syntax::Stmt::Let { expr, .. } | syntax::Stmt::Return { expr, .. } => {
-            validate_const_function_expr(function, expr, functions)
-        }
-        syntax::Stmt::If {
-            cond,
-            then_block,
-            else_block,
-            ..
-        } => {
-            validate_const_function_expr(function, cond, functions)?;
-            for stmt in then_block {
-                validate_const_function_stmt(function, stmt, functions)?;
-            }
-            if let Some(else_block) = else_block {
-                for stmt in else_block {
-                    validate_const_function_stmt(function, stmt, functions)?;
-                }
-            }
-            Ok(())
-        }
-        _ => Err(Diagnostic::new(
-            "type",
-            format!(
-                "const fn {:?} only supports let, if/else, and return statements in stage1",
-                function.name
-            ),
-        )
-        .with_span(stmt.line(), stmt.column())),
-    }
-}
-
-fn validate_const_function_expr(
-    function: &syntax::Function,
-    expr: &syntax::Expr,
-    functions: &HashMap<String, FunctionSig>,
-) -> Result<(), Diagnostic> {
-    match expr {
-        syntax::Expr::Literal(_) | syntax::Expr::VarRef { .. } => Ok(()),
-        syntax::Expr::BinaryAdd { lhs, rhs, .. } | syntax::Expr::BinaryCompare { lhs, rhs, .. } => {
-            validate_const_function_expr(function, lhs, functions)?;
-            validate_const_function_expr(function, rhs, functions)
-        }
-        syntax::Expr::Call { name, args, .. } => {
-            let Some(signature) = functions.get(name) else {
-                return Err(const_function_call_error(function, name, expr));
-            };
-            if !signature.is_const || signature.is_extern {
-                return Err(const_function_call_error(function, name, expr));
-            }
-            for arg in args {
-                validate_const_function_expr(function, arg, functions)?;
-            }
-            Ok(())
-        }
-        _ => Err(Diagnostic::new(
-            "type",
-            format!(
-                "const fn {:?} only supports literals, variables, arithmetic/comparison expressions, and calls to other const fn in stage1",
-                function.name
-            ),
-        )
-        .with_span(expr.line(), expr.column())),
-    }
-}
-
-fn const_function_call_error(
-    function: &syntax::Function,
-    callee: &str,
-    expr: &syntax::Expr,
-) -> Diagnostic {
-    Diagnostic::new(
-        "type",
-        format!(
-            "const fn {:?} can only call other const fn; {callee:?} is a host runtime or non-const call",
-            function.name
-        ),
-    )
-    .with_span(expr.line(), expr.column())
 }
 
 fn lower_function(
